workers/parser/requirements.txt
pytest==8.4.2
pypdf==5.9.0
python-docx==1.2.0
python-pptx==1.0.2
